MQTT - Alexander Lozada - David Felipe Rojas/requirements.txt
paho-mqtt
requests
python-dotenv
pymongo
